+++ README.md
@@ -1,11 +1,5 @@
-
 # NY Times App
-
 A sample Android project to apply latest technologies and architectures in Android.
-
-
-
-
 
 ## Tech Stack
 * Clean architecture (Oninon architecture).
