+++ app/src/androidTest/java/com/mahmoudelshamy/nytimes/core/HiltExtensions.kt
@@ -7,6 +7,7 @@ import androidx.fragment.app.Fragment
 import androidx.test.core.app.ActivityScenario
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.espresso.core.internal.deps.dagger.internal.Preconditions
+import com.mahmoudelshamy.nytimes.features.common.presentation.MainActivity
 
 inline fun <reified T : Fragment> launchFragmentInHiltContainer(
     fragmentArgs: Bundle? = null,
