+++ app/src/main/java/com/mahmoudelshamy/nytimes/features/articles/presentation/articlesList/ArticlesListFragment.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import androidx.fragment.app.viewModels
 import com.mahmoudelshamy.nytimes.R
-import com.mahmoudelshamy.nytimes.core.BaseFragment
+import com.mahmoudelshamy.nytimes.features.common.presentation.base.BaseFragment
 import com.mahmoudelshamy.nytimes.databinding.FragmentArticlesListBinding
 import com.mahmoudelshamy.nytimes.features.articles.domain.models.Article
 import dagger.hilt.android.AndroidEntryPoint
