+++ app/src/main/java/com/mahmoudelshamy/nytimes/features/common/presentation/base/BaseFragment.kt
@@ -1,4 +1,4 @@
-package com.mahmoudelshamy.nytimes.core
+package com.mahmoudelshamy.nytimes.features.common.presentation.base
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -10,7 +10,7 @@ import androidx.navigation.fragment.findNavController
 import androidx.navigation.navOptions
 import androidx.viewbinding.ViewBinding
 import com.mahmoudelshamy.nytimes.R
-import com.mahmoudelshamy.nytimes.common.NavigationCommand
+import com.mahmoudelshamy.nytimes.features.common.presentation.NavigationCommand
 
 abstract class BaseFragment<VB : ViewBinding, VM : BaseViewModel> : Fragment() {
     protected abstract val viewModel: VM
