+++ app/src/main/java/com/mahmoudelshamy/nytimes/features/common/presentation/base/BaseViewModel.kt
@@ -1,10 +1,10 @@
-package com.mahmoudelshamy.nytimes.core
+package com.mahmoudelshamy.nytimes.features.common.presentation.base
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import androidx.navigation.NavDirections
-import com.mahmoudelshamy.nytimes.common.NavigationCommand
-import com.mahmoudelshamy.nytimes.common.SingleLiveEvent
+import com.mahmoudelshamy.nytimes.features.common.presentation.NavigationCommand
+import com.mahmoudelshamy.nytimes.utils.SingleLiveEvent
 
 abstract class BaseViewModel : ViewModel() {
     private val _navigation = SingleLiveEvent<NavigationCommand>()
